backend/requirements.txt
#---Google Authentication---
google-auth==2.29.0
requests==2.31.0


# --- Core API ---
fastapi==0.110.0
uvicorn[standard]==0.29.0
pydantic==2.6.4
pydantic-settings==2.2.1

# --- Database (PostgreSQL & SQLAlchemy) ---
sqlalchemy==2.0.29
psycopg2-binary==2.9.9
alembic==1.13.1

# --- Security & Authentication (For User Login) ---
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.9

# --- Utilities ---
python-dotenv==1.0.1
